Add files via upload
Issue no.- 408

A jump search or block search refers to a search algorithm for ordered lists. It works by first checking all items Lkm, where  m is the block size, until an item is found that is larger than the search key. To find the exact position of the search key in the list a linear search is performed on the sublist L[(k-1)m, km].

I have uploaded the code of jump search in cpp language.